+++ features/pages/flaw_detail_page.py
@@ -73,8 +73,8 @@ class FlawDetailPage(BasePage):
         "publicDateText": ("XPATH", "//span[text()='Public Date']"),
         "publicDateValue": ("XPATH", "(//span[@class='osim-editable-date-value form-control text-start form-control'])[2]"),
 
-        "assigneeText": ("XPATH", "//span[contains(text(), 'Assignee')]"),
-        "assignee": ("XPATH", "//span[contains(text(), 'Assignee')]"),
+        "ownerText": ("XPATH", "//span[contains(text(), 'Owner')]"),
+        "owner": ("XPATH", "//span[contains(text(), 'Owner')]"),
         "selfAssignBtn": ("XPATH", "//button[contains(text(), 'Self Assign')]"),
 
         "referenceCountLabel": ("XPATH", '//label[contains(text(), "References:")]'),
@@ -109,7 +109,7 @@ class FlawDetailPage(BasePage):
         "affects__impact": ("XPATH", "(//span[text()='Impact'])[2]"),
         "affectUpdateMsg": ("XPATH", "//div[text()='Affect Updated.']"),
         "affectSaveMsg": ("XPATH", "//div[contains(text(), 'Affect 1 of 1 Saved:')]"),
-        "affectFileTrackerBtn": ("XPATH", "(//button[contains(text(), 'File Tracker')])[1]"),
+        "affectFileTrackerBtn": ("XPATH", "(//button[contains(text(), 'File Tracker')])[1]"), # TODO: update feature to use new tracker filing interface
         "affectDeleteTips": ("XPATH", "//h5[contains(text(), 'Affected Offerings To Be Deleted')]"),
         "affectDeleteMsg": ("XPATH", "//div[text()='Affect Deleted.']"),
         "affectRecoverBtn": ("XPATH", "//button[@title='Recover']"),