+++ features/utils.py
@@ -183,6 +183,6 @@ def get_osidb_token():
 
 def is_sorted(l, order):
     if order == 'desc':
-        return all(l[i] >= l[i+1] for i in range(len(l) - 1))
+        return l == sorted(l, reverse=True)
     else:
-        return all(l[i] <= l[i+1] for i in range(len(l) - 1))
+        return l == sorted(l)