+++ src/components/__tests__/FlawSearchView.spec.ts
@@ -199,4 +199,24 @@ describe('FlawSearchView', () => {
     expect(useRouter().replace.mock.calls[1][0])
       .toStrictEqual({ query: { query: 'search', requires_cve_description: 'REQUESTED' } });
   });
+
+  it('should call loadFlaws with major_incident_state on search', async () => {
+    await flushPromises();
+    expect(useFlawsFetching().loadFlaws).toHaveBeenCalledOnce();
+    const selectDropdown = wrapper.find('select.form-select.search-facet-field');
+    const dropdownOption = selectDropdown
+      .findAll('option')
+      .filter(option => option.element.value === 'major_incident_state')[0];
+    await selectDropdown.setValue(dropdownOption.element.value);
+    await selectDropdown.trigger('change');
+    const valueDropdown = wrapper.findAll('select.form-select')[1];
+    await valueDropdown.setValue(valueDropdown.findAll('option')[1].element.value);
+    const searchButton = wrapper.find('button[type="submit"]');
+    expect(searchButton.exists()).toBeTruthy();
+    await searchButton.trigger('submit');
+    await flushPromises();
+    expect(useRouter().replace).toHaveBeenCalled();
+    expect(useRouter().replace.mock.calls[2][0])
+      .toStrictEqual({ query: { query: 'search', major_incident_state: 'REQUESTED' } });
+  });
 });