+++ src/main.ts
@@ -12,8 +12,6 @@ import 'bootstrap/scss/bootstrap.scss';
 import 'bootstrap-icons/font/bootstrap-icons.scss';
 import './scss/index.scss';
 import './assets/main.css';
-
-import './services/service-worker-client';
 import 'bootstrap/js/index.esm.js';
 
 const app = createApp(App);